move manifest and builders to the base module
