+++ wren-core/core/Cargo.toml
@@ -33,3 +33,4 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 serde_with = { workspace = true }
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "macros"] }
+wren-core-base = { workspace = true }
